stats refactor

+++ alphasim/util.py
@@ -5,13 +5,13 @@ def like(source):
 
     match type(source):
         case pd.DataFrame:
-            copied = pd.DataFrame(np.zeros(source.shape))
+            copied = pd.DataFrame(np.zeros(source.shape, dtype=np.float64))
             copied.index = source.index
             copied.columns = source.columns
         case pd.Series:
-            copied = pd.Series(np.zeros(source.shape))
+            copied = pd.Series(np.zeros(source.shape, dtype=np.float64))
             copied.index = source.index
         case _:
-            copied = np.zeros(source.shape)
+            copied = np.zeros(source.shape, dtype=np.float64)
 
     return copied
